split runtime core api into focused submodules
