+++ crates/jazz-tools/src/runtime_core.rs
@@ -316,743 +316,6 @@ impl<S: Storage, Sch: Scheduler, Sy: SyncSender> RuntimeCore<S, Sch, Sy> {
         info!(object_id = %id, "persisted schema to catalogue");
         id
     }
-
-    // =========================================================================
-    // Tick Methods
-    // =========================================================================
-
-    /// Synchronous tick - processes managers, fulfills completed queries.
-    ///
-    /// Schedules batched_tick if there are outbound messages.
-    ///
-    /// Call this after any mutation operation (insert, update, delete, etc.)
-    /// to process the change and schedule any required I/O.
-    pub fn immediate_tick(&mut self) -> TickOutput {
-        let _span = trace_span!("immediate_tick", tier = self.tier_label).entered();
-
-        // 1. Process logical updates (sync, subscriptions)
-        self.schema_manager.process(&mut self.storage);
-
-        // 2. Second process() handles deferred query subscriptions that couldn't
-        //    compile on first pass (schema wasn't available yet, e.g. catalogue
-        //    was just processed and made the schema available).
-        self.schema_manager.process(&mut self.storage);
-
-        // 3. Collect subscription updates
-        let subscription_updates = self.schema_manager.query_manager_mut().take_updates();
-        let subscription_failures = self
-            .schema_manager
-            .query_manager_mut()
-            .take_failed_subscriptions();
-
-        // Track one-shot queries that completed this tick
-        let mut completed_one_shots: Vec<SubscriptionHandle> = Vec::new();
-        let mut failed_one_shots: Vec<SubscriptionHandle> = Vec::new();
-        let mut callbacks_fired: u64 = 0;
-
-        // 3. Call subscription callbacks AND handle one-shot queries
-        for update in &subscription_updates {
-            if let Some(&handle) = self.subscription_reverse.get(&update.subscription_id) {
-                // Check if this is a one-shot query
-                if let Some(pending) = self.pending_one_shot_queries.get_mut(&handle) {
-                    // First callback = graph settled, fulfill the future
-                    if let Some(sender) = pending.sender.take() {
-                        // Decode rows using the query's output descriptor
-                        let results: Vec<(ObjectId, Vec<Value>)> = update
-                            .delta
-                            .added
-                            .iter()
-                            .filter_map(|row| {
-                                decode_row(&update.descriptor, &row.data)
-                                    .ok()
-                                    .map(|values| (row.id, values))
-                            })
-                            .collect();
-                        let _ = sender.send(Ok(results));
-                    }
-                    // Mark for cleanup (unsubscribe happens after loop)
-                    completed_one_shots.push(handle);
-                } else if let Some(state) = self.subscriptions.get(&handle) {
-                    // Regular subscription - call callback
-                    let delta = SubscriptionDelta {
-                        handle,
-                        delta: update.delta.clone(),
-                        descriptor: update.descriptor.clone(),
-                    };
-                    (state.callback)(delta);
-                    callbacks_fired += 1;
-                }
-            }
-        }
-        tracing::debug!(callbacks_fired, "subscription callbacks fired this tick");
-
-        for failure in &subscription_failures {
-            if let Some(&handle) = self.subscription_reverse.get(&failure.subscription_id) {
-                if let Some(pending) = self.pending_one_shot_queries.get_mut(&handle) {
-                    if let Some(sender) = pending.sender.take() {
-                        let _ = sender.send(Err(RuntimeError::QueryError(format!(
-                            "query subscription {} failed during schema recompile: {}",
-                            failure.subscription_id.0, failure.reason
-                        ))));
-                    }
-                    failed_one_shots.push(handle);
-                } else if self.subscriptions.remove(&handle).is_some() {
-                    self.subscription_reverse.remove(&failure.subscription_id);
-                    tracing::error!(
-                        handle = handle.0,
-                        sub_id = failure.subscription_id.0,
-                        error = %failure.reason,
-                        "subscription failed during schema recompile and was dropped"
-                    );
-                }
-            } else {
-                tracing::error!(
-                    sub_id = failure.subscription_id.0,
-                    error = %failure.reason,
-                    "subscription failed during schema recompile and was dropped"
-                );
-            }
-        }
-
-        // 2b. Cleanup completed one-shot queries
-        for handle in completed_one_shots {
-            if let Some(pending) = self.pending_one_shot_queries.remove(&handle) {
-                // Unsubscribe from the underlying subscription
-                self.schema_manager
-                    .query_manager_mut()
-                    .unsubscribe_with_sync(pending.subscription_id);
-                self.subscription_reverse.remove(&pending.subscription_id);
-            }
-        }
-
-        // 2c. Cleanup failed one-shot queries.
-        // The underlying subscriptions were already removed by QueryManager.
-        for handle in failed_one_shots {
-            if let Some(pending) = self.pending_one_shot_queries.remove(&handle) {
-                self.subscription_reverse.remove(&pending.subscription_id);
-            }
-        }
-
-        // 3b. Process received persistence acks — resolve matching watchers
-        let received_acks = self
-            .schema_manager
-            .query_manager_mut()
-            .sync_manager_mut()
-            .take_received_acks();
-        for (commit_id, acked_tier) in received_acks {
-            if let Some(watchers) = self.ack_watchers.remove(&commit_id) {
-                let mut remaining = Vec::new();
-                for (requested_tier, sender) in watchers {
-                    if acked_tier >= requested_tier {
-                        tracing::debug!(
-                            ?commit_id,
-                            ?acked_tier,
-                            ?requested_tier,
-                            "ack watcher resolved"
-                        );
-                        let _ = sender.send(());
-                    } else {
-                        remaining.push((requested_tier, sender));
-                    }
-                }
-                if !remaining.is_empty() {
-                    self.ack_watchers.insert(commit_id, remaining);
-                }
-            }
-        }
-
-        // 4. Schedule batched_tick if outbound messages exist
-        if self.has_outbound() {
-            self.scheduler.schedule_batched_tick();
-        }
-
-        TickOutput {
-            subscription_updates,
-        }
-    }
-
-    /// Batched tick - handles all I/O, then processes parked messages.
-    ///
-    /// Called by the platform when the scheduled tick fires. This:
-    /// 1. Sends all outgoing sync messages via SyncSender
-    /// 2. Processes parked sync messages
-    ///
-    /// Each step is followed by an immediate_tick to process results.
-    pub fn batched_tick(&mut self) {
-        let _span = debug_span!("batched_tick", tier = self.tier_label).entered();
-
-        // 1. Send all outgoing sync messages
-        let outbox = self
-            .schema_manager
-            .query_manager_mut()
-            .sync_manager_mut()
-            .take_outbox();
-        if !outbox.is_empty() {
-            debug!(count = outbox.len(), "flushing outbox");
-        }
-        for msg in outbox {
-            self.sync_sender.send_sync_message(msg);
-        }
-
-        // 2. Process parked sync messages
-        self.handle_sync_messages();
-
-        // 3. Flush any new outbox entries generated by processing.
-        // The scheduler's debounce prevents immediate_tick() from scheduling
-        // another batched_tick while we're inside one, so we must flush here.
-        let outbox = self
-            .schema_manager
-            .query_manager_mut()
-            .sync_manager_mut()
-            .take_outbox();
-        if !outbox.is_empty() {
-            debug!(count = outbox.len(), "flushing post-process outbox");
-        }
-        for msg in outbox {
-            self.sync_sender.send_sync_message(msg);
-        }
-
-        // Flush the storage durability barrier so writes survive a hard kill (tab close, crash).
-        {
-            let _span = tracing::debug_span!("flush_wal").entered();
-            self.storage.flush_wal();
-        }
-    }
-
-    /// Apply parked sync messages and tick.
-    fn handle_sync_messages(&mut self) {
-        let messages = std::mem::take(&mut self.parked_sync_messages);
-        let mut applied_messages = 0usize;
-
-        if !messages.is_empty() {
-            debug!(
-                count = messages.len(),
-                "processing parked unsequenced sync messages"
-            );
-        }
-        for msg in messages {
-            self.push_sync_inbox(msg);
-            applied_messages += 1;
-        }
-
-        let server_ids: Vec<ServerId> = self
-            .parked_sync_messages_by_server_seq
-            .keys()
-            .copied()
-            .collect();
-        for server_id in server_ids {
-            let mut next_expected = *self.next_expected_server_seq.get(&server_id).unwrap_or(&1);
-            let mut ready_messages = Vec::new();
-            let mut remove_buffer = false;
-            if let Some(buffered) = self.parked_sync_messages_by_server_seq.get_mut(&server_id) {
-                while let Some(msg) = buffered.remove(&next_expected) {
-                    ready_messages.push(msg);
-                    next_expected += 1;
-                }
-
-                if buffered.is_empty() {
-                    remove_buffer = true;
-                }
-            }
-            for msg in ready_messages {
-                self.push_sync_inbox(msg);
-                applied_messages += 1;
-            }
-            self.next_expected_server_seq
-                .insert(server_id, next_expected);
-            if remove_buffer {
-                self.parked_sync_messages_by_server_seq.remove(&server_id);
-            }
-        }
-
-        if applied_messages > 0 {
-            debug!(count = applied_messages, "applied parked sync messages");
-            self.immediate_tick();
-        }
-    }
-
-    /// Check if there are outbound messages requiring a batched_tick.
-    pub fn has_outbound(&self) -> bool {
-        !self
-            .schema_manager
-            .query_manager()
-            .sync_manager()
-            .outbox()
-            .is_empty()
-    }
-
-    /// Park a sync message for processing in next batched_tick.
-    pub fn park_sync_message(&mut self, message: InboxEntry) {
-        trace!(source = ?message.source, payload = message.payload.variant_name(), "parking sync message");
-        self.parked_sync_messages.push(message);
-        self.scheduler.schedule_batched_tick();
-    }
-
-    /// Park a sequenced sync message for in-order processing in next batched_tick.
-    pub fn park_sync_message_with_sequence(&mut self, message: InboxEntry, sequence: u64) {
-        match message.source {
-            crate::sync_manager::Source::Server(server_id) => {
-                let next_expected = self
-                    .next_expected_server_seq
-                    .entry(server_id)
-                    .or_insert(sequence);
-                if sequence < *next_expected {
-                    trace!(
-                        ?server_id,
-                        sequence,
-                        next_expected = *next_expected,
-                        "dropping already-applied sequenced sync message"
-                    );
-                    return;
-                }
-
-                self.parked_sync_messages_by_server_seq
-                    .entry(server_id)
-                    .or_default()
-                    .insert(sequence, message);
-                self.scheduler.schedule_batched_tick();
-            }
-            _ => self.park_sync_message(message),
-        }
-    }
-
-    /// Set the next expected sequenced message for a server stream.
-    pub fn set_next_expected_server_sequence(&mut self, server_id: ServerId, next_sequence: u64) {
-        let next_sequence = next_sequence.max(1);
-        self.next_expected_server_seq
-            .insert(server_id, next_sequence);
-        if let Some(buffered) = self.parked_sync_messages_by_server_seq.get_mut(&server_id) {
-            buffered.retain(|seq, _| *seq >= next_sequence);
-        }
-    }
-
-    // =========================================================================
-    // Subscriptions
-    // =========================================================================
-
-    /// Subscribe to a query with a callback.
-    #[cfg(not(target_arch = "wasm32"))]
-    pub fn subscribe<F>(
-        &mut self,
-        query: Query,
-        callback: F,
-        session: Option<Session>,
-    ) -> Result<SubscriptionHandle, RuntimeError>
-    where
-        F: Fn(SubscriptionDelta) + Send + 'static,
-    {
-        self.subscribe_impl(query, Box::new(callback), session, None)
-    }
-
-    /// Subscribe to a query with a callback (WASM version - no Send required).
-    #[cfg(target_arch = "wasm32")]
-    pub fn subscribe<F>(
-        &mut self,
-        query: Query,
-        callback: F,
-        session: Option<Session>,
-    ) -> Result<SubscriptionHandle, RuntimeError>
-    where
-        F: Fn(SubscriptionDelta) + 'static,
-    {
-        self.subscribe_impl(query, Box::new(callback), session, None)
-    }
-
-    /// Subscribe with optional settled tier.
-    #[cfg(not(target_arch = "wasm32"))]
-    pub fn subscribe_with_settled_tier<F>(
-        &mut self,
-        query: Query,
-        callback: F,
-        session: Option<Session>,
-        settled_tier: Option<PersistenceTier>,
-    ) -> Result<SubscriptionHandle, RuntimeError>
-    where
-        F: Fn(SubscriptionDelta) + Send + 'static,
-    {
-        self.subscribe_impl(query, Box::new(callback), session, settled_tier)
-    }
-
-    /// Subscribe with settled tier (WASM version - no Send required).
-    #[cfg(target_arch = "wasm32")]
-    pub fn subscribe_with_settled_tier<F>(
-        &mut self,
-        query: Query,
-        callback: F,
-        session: Option<Session>,
-        settled_tier: Option<PersistenceTier>,
-    ) -> Result<SubscriptionHandle, RuntimeError>
-    where
-        F: Fn(SubscriptionDelta) + 'static,
-    {
-        self.subscribe_impl(query, Box::new(callback), session, settled_tier)
-    }
-
-    /// Internal subscribe implementation.
-    fn subscribe_impl(
-        &mut self,
-        query: Query,
-        callback: SubscriptionCallback,
-        session: Option<Session>,
-        settled_tier: Option<PersistenceTier>,
-    ) -> Result<SubscriptionHandle, RuntimeError> {
-        let _span = debug_span!("subscribe", table = query.table.as_str()).entered();
-        let query_sub_id = self
-            .schema_manager
-            .query_manager_mut()
-            .subscribe_with_sync(query, session, settled_tier)
-            .map_err(|e| RuntimeError::QueryError(format!("{:?}", e)))?;
-
-        let handle = SubscriptionHandle(self.next_subscription_handle);
-        self.next_subscription_handle += 1;
-        debug!(handle = handle.0, sub_id = query_sub_id.0, "subscribed");
-
-        self.subscriptions.insert(
-            handle,
-            SubscriptionState {
-                query_sub_id,
-                callback,
-            },
-        );
-        self.subscription_reverse.insert(query_sub_id, handle);
-
-        self.immediate_tick();
-        Ok(handle)
-    }
-
-    /// Unsubscribe from a query.
-    pub fn unsubscribe(&mut self, handle: SubscriptionHandle) {
-        if let Some(state) = self.subscriptions.remove(&handle) {
-            self.subscription_reverse.remove(&state.query_sub_id);
-            self.schema_manager
-                .query_manager_mut()
-                .unsubscribe_with_sync(state.query_sub_id);
-        }
-    }
-
-    /// Subscribe with explicit schema context (for server use).
-    pub fn subscribe_with_schema_context(
-        &mut self,
-        query: Query,
-        schema_context: &crate::schema_manager::QuerySchemaContext,
-        session: Option<Session>,
-    ) -> Result<crate::sync_manager::QueryId, RuntimeError> {
-        let query_sub_id = self
-            .schema_manager
-            .subscribe_with_schema_context(query, schema_context, session)
-            .map_err(|e| RuntimeError::QueryError(format!("{:?}", e)))?;
-
-        self.immediate_tick();
-        Ok(crate::sync_manager::QueryId(query_sub_id.0))
-    }
-
-    // =========================================================================
-    // Queries
-    // =========================================================================
-
-    /// Execute a one-shot query, optionally waiting for a settled tier.
-    pub fn query(
-        &mut self,
-        query: Query,
-        session: Option<Session>,
-        settled_tier: Option<PersistenceTier>,
-    ) -> QueryFuture {
-        let _span = debug_span!("query", table = query.table.as_str(), ?settled_tier).entered();
-        let (sender, receiver) = oneshot::channel();
-
-        let sub_id = match self.schema_manager.query_manager_mut().subscribe_with_sync(
-            query,
-            session,
-            settled_tier,
-        ) {
-            Ok(id) => id,
-            Err(e) => {
-                let _ = sender.send(Err(RuntimeError::QueryError(format!("{:?}", e))));
-                return QueryFuture::new(receiver);
-            }
-        };
-
-        let handle = SubscriptionHandle(self.next_subscription_handle);
-        self.next_subscription_handle += 1;
-
-        self.pending_one_shot_queries.insert(
-            handle,
-            PendingOneShotQuery {
-                subscription_id: sub_id,
-                sender: Some(sender),
-            },
-        );
-        self.subscription_reverse.insert(sub_id, handle);
-
-        self.immediate_tick();
-        QueryFuture::new(receiver)
-    }
-
-    // =========================================================================
-    // CRUD Operations
-    // =========================================================================
-
-    /// Insert a row into a table.
-    pub fn insert(
-        &mut self,
-        table: &str,
-        values: Vec<Value>,
-        session: Option<&Session>,
-    ) -> Result<ObjectId, RuntimeError> {
-        let _span = debug_span!("insert", table).entered();
-        let result = self
-            .schema_manager
-            .insert_with_session(&mut self.storage, table, &values, session)
-            .map_err(|e| RuntimeError::WriteError(format!("{:?}", e)))?;
-        debug!(object_id = %result.row_id, "inserted");
-        self.immediate_tick();
-        Ok(result.row_id)
-    }
-
-    /// Update a row (partial update by column name).
-    pub fn update(
-        &mut self,
-        object_id: ObjectId,
-        values: Vec<(String, Value)>,
-        session: Option<&Session>,
-    ) -> Result<(), RuntimeError> {
-        let _span = debug_span!("update", %object_id).entered();
-        let (table, mut current_values) = self
-            .schema_manager
-            .query_manager_mut()
-            .get_row(object_id)
-            .ok_or(RuntimeError::NotFound)?;
-
-        let schema = self.schema_manager.current_schema();
-        let table_name = TableName::new(&table);
-        let table_schema = schema
-            .get(&table_name)
-            .ok_or_else(|| RuntimeError::WriteError("Table not found".to_string()))?;
-
-        for (col_name, new_value) in values {
-            if let Some(idx) = table_schema.descriptor.column_index(&col_name) {
-                current_values[idx] = new_value;
-            } else {
-                return Err(RuntimeError::WriteError(format!(
-                    "Column '{}' not found",
-                    col_name
-                )));
-            }
-        }
-
-        self.schema_manager
-            .query_manager_mut()
-            .update_with_session(&mut self.storage, object_id, &current_values, session)
-            .map_err(|e| RuntimeError::WriteError(format!("{:?}", e)))?;
-
-        self.immediate_tick();
-        Ok(())
-    }
-
-    /// Delete a row.
-    pub fn delete(
-        &mut self,
-        object_id: ObjectId,
-        session: Option<&Session>,
-    ) -> Result<(), RuntimeError> {
-        let _span = debug_span!("delete", %object_id).entered();
-        self.schema_manager
-            .query_manager_mut()
-            .delete_with_session(&mut self.storage, object_id, session)
-            .map_err(|e| RuntimeError::WriteError(format!("{:?}", e)))?;
-        debug!("deleted");
-        self.immediate_tick();
-        Ok(())
-    }
-
-    // =========================================================================
-    // Persisted CRUD Operations
-    // =========================================================================
-
-    /// Insert a row and return a receiver that resolves when the requested
-    /// persistence tier (or higher) acknowledges.
-    pub fn insert_persisted(
-        &mut self,
-        table: &str,
-        values: Vec<Value>,
-        session: Option<&Session>,
-        tier: PersistenceTier,
-    ) -> Result<(ObjectId, oneshot::Receiver<()>), RuntimeError> {
-        let result = self
-            .schema_manager
-            .insert_with_session(&mut self.storage, table, &values, session)
-            .map_err(|e| RuntimeError::WriteError(format!("{:?}", e)))?;
-
-        let (sender, receiver) = oneshot::channel();
-        self.ack_watchers
-            .entry(result.row_commit_id)
-            .or_default()
-            .push((tier, sender));
-
-        self.immediate_tick();
-        Ok((result.row_id, receiver))
-    }
-
-    /// Update a row and return a receiver that resolves when the requested
-    /// persistence tier (or higher) acknowledges.
-    pub fn update_persisted(
-        &mut self,
-        object_id: ObjectId,
-        values: Vec<(String, Value)>,
-        session: Option<&Session>,
-        tier: PersistenceTier,
-    ) -> Result<oneshot::Receiver<()>, RuntimeError> {
-        let (table, mut current_values) = self
-            .schema_manager
-            .query_manager_mut()
-            .get_row(object_id)
-            .ok_or(RuntimeError::NotFound)?;
-
-        let schema = self.schema_manager.current_schema();
-        let table_name = TableName::new(&table);
-        let table_schema = schema
-            .get(&table_name)
-            .ok_or_else(|| RuntimeError::WriteError("Table not found".to_string()))?;
-
-        for (col_name, new_value) in values {
-            if let Some(idx) = table_schema.descriptor.column_index(&col_name) {
-                current_values[idx] = new_value;
-            } else {
-                return Err(RuntimeError::WriteError(format!(
-                    "Column '{}' not found",
-                    col_name
-                )));
-            }
-        }
-
-        let commit_id = self
-            .schema_manager
-            .query_manager_mut()
-            .update_with_session(&mut self.storage, object_id, &current_values, session)
-            .map_err(|e| RuntimeError::WriteError(format!("{:?}", e)))?;
-
-        let (sender, receiver) = oneshot::channel();
-        self.ack_watchers
-            .entry(commit_id)
-            .or_default()
-            .push((tier, sender));
-
-        self.immediate_tick();
-        Ok(receiver)
-    }
-
-    /// Delete a row and return a receiver that resolves when the requested
-    /// persistence tier (or higher) acknowledges.
-    pub fn delete_persisted(
-        &mut self,
-        object_id: ObjectId,
-        session: Option<&Session>,
-        tier: PersistenceTier,
-    ) -> Result<oneshot::Receiver<()>, RuntimeError> {
-        let handle = self
-            .schema_manager
-            .query_manager_mut()
-            .delete_with_session(&mut self.storage, object_id, session)
-            .map_err(|e| RuntimeError::WriteError(format!("{:?}", e)))?;
-
-        let (sender, receiver) = oneshot::channel();
-        self.ack_watchers
-            .entry(handle.delete_commit_id)
-            .or_default()
-            .push((tier, sender));
-
-        self.immediate_tick();
-        Ok(receiver)
-    }
-
-    // =========================================================================
-    // Sync Operations
-    // =========================================================================
-
-    /// Push a sync message to the inbox (from network).
-    pub fn push_sync_inbox(&mut self, entry: InboxEntry) {
-        self.schema_manager
-            .query_manager_mut()
-            .sync_manager_mut()
-            .push_inbox(entry);
-    }
-
-    /// Add a server connection.
-    pub fn add_server(&mut self, server_id: ServerId) {
-        info!(%server_id, "adding server");
-        self.schema_manager
-            .query_manager_mut()
-            .add_server(server_id);
-        self.immediate_tick();
-    }
-
-    /// Remove a server connection.
-    pub fn remove_server(&mut self, server_id: ServerId) {
-        self.schema_manager
-            .query_manager_mut()
-            .sync_manager_mut()
-            .remove_server(server_id);
-    }
-
-    /// Add a client connection.
-    pub fn add_client(&mut self, client_id: ClientId, session: Option<Session>) {
-        info!(%client_id, has_session = session.is_some(), "adding client");
-        let sm = self.schema_manager.query_manager_mut().sync_manager_mut();
-        sm.add_client(client_id);
-        if let Some(s) = session {
-            sm.set_client_session(client_id, s);
-        }
-        self.immediate_tick();
-    }
-
-    /// Ensure a client exists with the given session.
-    ///
-    /// If the client already exists, updates the session. This is idempotent —
-    /// calling with the same session is a no-op. Calling with a new session
-    /// updates it in place without resetting the client's role or other state.
-    ///
-    /// A session is always required — callers must authenticate before
-    /// registering a client.
-    pub fn ensure_client_with_session(&mut self, client_id: ClientId, session: Session) {
-        let sm = self.schema_manager.query_manager_mut().sync_manager_mut();
-        if sm.get_client(client_id).is_some() {
-            sm.set_client_session(client_id, session);
-        } else {
-            sm.add_client(client_id);
-            sm.set_client_session(client_id, session);
-            self.immediate_tick();
-        }
-    }
-
-    /// Remove a client connection.
-    pub fn remove_client(&mut self, client_id: ClientId) {
-        self.schema_manager
-            .query_manager_mut()
-            .sync_manager_mut()
-            .remove_client(client_id);
-    }
-
-    /// Promote a client to Admin role (full access, no ReBAC).
-    pub fn set_client_admin(&mut self, client_id: ClientId) {
-        use crate::sync_manager::ClientRole;
-        self.schema_manager
-            .query_manager_mut()
-            .sync_manager_mut()
-            .set_client_role(client_id, ClientRole::Admin);
-    }
-
-    /// Set a client's role.
-    pub fn set_client_role_by_name(
-        &mut self,
-        client_id: ClientId,
-        role: crate::sync_manager::ClientRole,
-    ) {
-        self.schema_manager
-            .query_manager_mut()
-            .sync_manager_mut()
-            .set_client_role(client_id, role);
-    }
-
     // =========================================================================
     // Schema/State Access
     // =========================================================================
@@ -1072,6 +335,11 @@ impl<S: Storage, Sch: Scheduler, Sy: SyncSender> RuntimeCore<S, Sch, Sy> {
         &self.schema_manager
     }
 }
+
+mod subscriptions;
+mod sync;
+mod ticks;
+mod writes;
 
 #[cfg(test)]
 mod tests;
