+++ crates/jazz-tools/src/runtime_core/sync.rs
@@ -1,0 +1,91 @@
+use super::*;
+
+impl<S: Storage, Sch: Scheduler, Sy: SyncSender> RuntimeCore<S, Sch, Sy> {
+    // =========================================================================
+    // Sync Operations
+    // =========================================================================
+
+    /// Push a sync message to the inbox (from network).
+    pub fn push_sync_inbox(&mut self, entry: InboxEntry) {
+        self.schema_manager
+            .query_manager_mut()
+            .sync_manager_mut()
+            .push_inbox(entry);
+    }
+
+    /// Add a server connection.
+    pub fn add_server(&mut self, server_id: ServerId) {
+        info!(%server_id, "adding server");
+        self.schema_manager
+            .query_manager_mut()
+            .add_server(server_id);
+        self.immediate_tick();
+    }
+
+    /// Remove a server connection.
+    pub fn remove_server(&mut self, server_id: ServerId) {
+        self.schema_manager
+            .query_manager_mut()
+            .sync_manager_mut()
+            .remove_server(server_id);
+    }
+
+    /// Add a client connection.
+    pub fn add_client(&mut self, client_id: ClientId, session: Option<Session>) {
+        info!(%client_id, has_session = session.is_some(), "adding client");
+        let sm = self.schema_manager.query_manager_mut().sync_manager_mut();
+        sm.add_client(client_id);
+        if let Some(s) = session {
+            sm.set_client_session(client_id, s);
+        }
+        self.immediate_tick();
+    }
+
+    /// Ensure a client exists with the given session.
+    ///
+    /// If the client already exists, updates the session. This is idempotent —
+    /// calling with the same session is a no-op. Calling with a new session
+    /// updates it in place without resetting the client's role or other state.
+    ///
+    /// A session is always required — callers must authenticate before
+    /// registering a client.
+    pub fn ensure_client_with_session(&mut self, client_id: ClientId, session: Session) {
+        let sm = self.schema_manager.query_manager_mut().sync_manager_mut();
+        if sm.get_client(client_id).is_some() {
+            sm.set_client_session(client_id, session);
+        } else {
+            sm.add_client(client_id);
+            sm.set_client_session(client_id, session);
+            self.immediate_tick();
+        }
+    }
+
+    /// Remove a client connection.
+    pub fn remove_client(&mut self, client_id: ClientId) {
+        self.schema_manager
+            .query_manager_mut()
+            .sync_manager_mut()
+            .remove_client(client_id);
+    }
+
+    /// Promote a client to Admin role (full access, no ReBAC).
+    pub fn set_client_admin(&mut self, client_id: ClientId) {
+        use crate::sync_manager::ClientRole;
+        self.schema_manager
+            .query_manager_mut()
+            .sync_manager_mut()
+            .set_client_role(client_id, ClientRole::Admin);
+    }
+
+    /// Set a client's role.
+    pub fn set_client_role_by_name(
+        &mut self,
+        client_id: ClientId,
+        role: crate::sync_manager::ClientRole,
+    ) {
+        self.schema_manager
+            .query_manager_mut()
+            .sync_manager_mut()
+            .set_client_role(client_id, role);
+    }
+}
